+++ integrations/dotnet/aspire/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @scalar/aspire
 
+## 0.8.15
+
+### Patch Changes
+
+#### Updated Dependencies
+
+- **@scalar/api-reference@1.43.3**
+
+## 0.8.14
+
+### Patch Changes
+
+- [#7698](https://github.com/scalar/scalar/pull/7698): feat: update Docker Hub description
+
+#### Updated Dependencies
+
+- **@scalar/api-reference@1.43.2**
+
 ## 0.8.13
 
 ### Patch Changes