+++ packages/mock-server/docker/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @scalar/mock-server-docker
 
+## 0.1.6
+
+### Patch Changes
+
+#### Updated Dependencies
+
+- **@scalar/mock-server@0.8.10**
+
+- **@scalar/hono-api-reference@0.9.32**
+
+## 0.1.5
+
+### Patch Changes
+
+- [#7698](https://github.com/scalar/scalar/pull/7698): feat: update Docker Hub description
+
+#### Updated Dependencies
+
+- **@scalar/hono-api-reference@0.9.32**
+
 ## 0.1.4
 
 ### Patch Changes