+++ packages/openapi-to-markdown/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @scalar/openapi-to-markdown
 
+## 0.3.18
+
+### Patch Changes
+
+#### Updated Dependencies
+
+- **@scalar/openapi-parser@0.23.12**
+  - [#7719](https://github.com/scalar/scalar/pull/7719): fix: self-referencing causes infinite loop in `dereference`
+
+- **@scalar/oas-utils@0.6.15**
+
+- **@scalar/components@0.16.18**
+
+## 0.3.17
+
+### Patch Changes
+
+#### Updated Dependencies
+
+- **@scalar/components@0.16.17**
+  - [#7680](https://github.com/scalar/scalar/pull/7680): feat(components): simplify nested sidebar items styles
+
 ## 0.3.16
 
 ### Patch Changes